refactor: Organize folder structures
