+++ src/App.js
@@ -1,9 +1,10 @@
 import "./App.css";
 import LoginPage from "../src/pages/LoginPage";
+import HomepyPage from './pages/homepyPage.js';
 import React, { useState } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import LoginCallback from "../src/components/LoginPage/LoginCallback";
-import Homepy from './homepy.js';
+import LoginCallback from "./components/LoginPage/LoginCallback.js";
+
 
 function App() {
   return (
@@ -13,7 +14,7 @@ function App() {
         <Route path="/" element={<LoginPage />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/login/auth/callback/google" element={<LoginCallback />} />
-        <Route path="/homepy" element={<Homepy />} />
+        <Route path="/homepy" element={<HomepyPage />} />
       </Routes>
     </BrowserRouter>
   );
