+++ src/components/HomepyPage/Album.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 
-import "../homepy-style.css";
-import "../homepy-styleguide.css";
+import "../../pages/homepy-style.css";
+import "../../pages/homepy-styleguide.css";
 
-import ph_plus_fill from "../assets/ph_plus-fill.svg";
+import ph_plus_fill from "../../assets/ph_plus-fill.svg";
 
 
 export default function Album() {
